ci: added UI tests back
